refactor: extract rag_clarification.py from rag_core (C-4, #563)
Move 4 clarification-path functions out of rag_core.py into a new leaf
module rag_clarification.py: clarification_answer,
make_context_clarification_result, metadata_clarification_answer,
make_metadata_clarification_result.

rag_clarification imports from rag_text_processing, rag_pipeline_presets,
rag_answer_schema, rag_answer, rag_verifier, rag_query, rag_tracing, and
stdlib — no rag_core dependency. rag_core re-exports all 4 names so
existing orchestration callers remain unchanged.

rag_core.py: 2,857 → 1,684 lines across the full C-track decomposition.
Late-imports eliminated: 0 remaining circular-import workarounds.
